Enable execution to scroll automatically to the bottom
This PR uses _process to auto scroll the execution scene to the bottom to keep up with the output.

Sorry about the previous accidental PR, using GitHub in the web messed me up. I will setup soon to commit via git CLI ssh.

Please consider this as one of my first submissions for this project. I hope to make many more contributions as I can see a lot of value in this program.